DP analysis graphs with DPJumboPlot
resolved large file merge errors

+++ BDT/crab_bdt_generator.py
@@ -14,6 +14,7 @@ if __name__ == "__main__":
     parser.add_argument("--outputDir")
     parser.add_argument("--outputFile")
     parser.add_argument("--trainingArgs")
+    parser.add_argument("--useCachedRepo", action="store_true")
     
     args = parser.parse_args()
     
@@ -24,32 +25,9 @@ if __name__ == "__main__":
     # script_content: script to run on condor
     
     script_content = (f"""
-# ##################### SETUP #####################
-
-ls
-
-echo "########################################"
-
-pwd
-ls
-
-# # Extract xxhash locally
-# wget https://github.com/Cyan4973/xxHash/archive/v0.8.2.tar.gz
-# tar xzf v0.8.2.tar.gz
-# cd xxHash-0.8.2
-# make
-# # Now set LD_LIBRARY_PATH in your job script:
-# export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$(pwd)
-
-# # Extract root locally
-# wget https://root.cern/download/root_v6.32.10.Linux-almalinux9.5-x86_64-gcc11.5.tar.gz
-# tar -xzvf root_v6.32.10.Linux-almalinux9.5-x86_64-gcc11.5.tar.gz
-# source root/bin/thisroot.sh
-
-# ##################### EXECUTION #####################
+echo "##################### SETUP #####################"
 
 source /cvmfs/cms.cern.ch/cmsset_default.sh
-
 cmsrel CMSSW_14_0_4
 
 # Extract input directories
@@ -57,25 +35,23 @@ cmsrel CMSSW_14_0_4
 tar -xzvf uncompiled.tar.gz
 tar -xzvf condor_input.tar.gz
 
-cp -r condor_input CMSSW_14_0_4/src/CMSAnalysis/BDT
 cp -r CMSAnalysis CMSSW_14_0_4/src
+cp -r condor_input CMSSW_14_0_4/src/CMSAnalysis/BDT
 
 cd CMSSW_14_0_4/src/CMSAnalysis
-
 cmsenv
+
+echo "################### COMPILING #####################"
+
+pwd
 
 # -j2 flag necessary to prevent exceeding memory quota (j4 > 2047, j8 maybe for 8192)
 scram b clean
-scram b -j8
+scram b -j
 
 cd BDT
 
-# echo "########################################"
-# printenv
-
-echo "########################################"
-find $CMSSW_BASE/src/CMSAnalysis/Modules/interface -name "LeptonJetMLStripModule.hh"
-echo $ROOT_INCLUDE_PATH
+echo "################### RUNNING ROOT #####################"
 
 root \'crab_BDT_MLTrain.C(\"condor_input/\", \"{args.outputDir}\", \"{args.outputFile}\", \"{args.trainingArgs}\")\'
 """
@@ -93,51 +69,62 @@ root \'crab_BDT_MLTrain.C(\"condor_input/\", \"{args.outputDir}\", \"{args.outpu
     jdl_filename = "crab_bdt_request.jdl"
     open(jdl_filename, "w").write(f"""                                         
 universe = vanilla
-request_memory = 8192
+request_memory = 32768
 request_cpus = 4
 Executable = {script_filename}
 should_transfer_files = YES
 when_to_transfer_output = ON_EXIT
 Output = condor_logs/bdt_$(Cluster)_$(Process).stdout
+Error = condor_logs/bdt_$(Cluster)_$(Process).stderr
 Log = condor_logs/bdt_$(Cluster)_$(Process).log
 transfer_input_files = uncompiled.tar.gz,condor_input.tar.gz
 transfer_output_files = {args.outputFile}.root,{args.outputFile}Rerun.root,TMVAClassification_BDT.class.C,TMVAClassification_BDT.weights.xml
 Queue 1
 """)
-    
-    # Error = condor_logs/bdt_$(Cluster)_$(Process).stderr
+    # request_cpus = 4 
     
     print("JDL file generated successfully")
     
     execute_filename = "execute_condor.sh"
+
     open(execute_filename, "w").write(f"""                                         
 #!/bin/bash
 
-# Update repo
+# clean up request files (if necessary)
+# rm uncompiled.tar.gz
+# rm condor_input.tar.gz
+# rm crab_bdt_request.jdl
+# rm execute_condor.sh
+# rm training_shell.sh
 
-mkdir /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo
-git clone git@github.com:IMSA-CMS/CMSAnalysis.git /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo
-tar -czf uncompiled.tar.gz /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo/CMSAnalysis
-rm -rf /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo
+# Check if repo update is required
+if [ "$1" != "useCachedRepo" ]; then
+    echo "Updating repository..."
+    mkdir -p /eos/uscms/store/user/{os.environ["USER"]}/condor_utils
+    git clone git@github.com:IMSA-CMS/CMSAnalysis.git /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo/CMSAnalysis
+    cd /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo/CMSAnalysis
+    git pull
+    
+    cd -
+    tar -czf uncompiled.tar.gz -C /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo CMSAnalysis
+    rm -rf /eos/uscms/store/user/{os.environ["USER"]}/condor_utils/temp_repo
+else
+    echo "Using cached repository (no git update)"
+fi
 
 # Fetch input files
-
 cp -r {args.input} ./condor_input
 tar -czf condor_input.tar.gz ./condor_input
 rm -rf ./condor_input
 
-condor_submit $1
+condor_submit $2
+    """)
 
-# clean up request files
+    # Pass "useCachedRepo" only if the flag is set
+    use_cached_flag = "useCachedRepo" if args.useCachedRepo else ""
 
-rm uncompiled.tar.gz
-rm condor_input.tar.gz
-rm crab_bdt_request.jdl
-rm execute_condor.sh
-rm training_shell.sh
-""")
-    
     submit = Popen(
-        ["bash", bdt_directory + execute_filename, jdl_filename], cwd=bdt_directory
+        ["bash", bdt_directory + execute_filename, use_cached_flag, jdl_filename],
+        cwd=bdt_directory
     )
     submit.wait()
